added Favorite route and add/delete feature for favorite photos

+++ src/Context.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState, useReducer } from 'react';
+import React, { createContext, useContext, useReducer } from 'react';
 import { CtxTypes, DetailAction, DetailState } from './Types';
 
 const ctx = createContext<CtxTypes | null>(null);
@@ -20,18 +20,19 @@ interface Props {
 const Context: React.FC<Props> = ({ children }) => {
   const [detailState, dispatchDetail] = useReducer(detailReducer, {} as DetailState);
 
-  const addToLocalStorage = (id: string, url: string) => {
+  const addToLocalStorage = () => {
     const favorites = JSON.parse(localStorage.getItem('favorites') ?? '[]');
-    const value = favorites.length ? [...favorites, {id, url}] : [{id, url}];
+    const alreadyExists = favorites.some((photo : DetailState) => photo.id === detailState.id)
+    if (alreadyExists) return;
+    const value = favorites.length ? [...favorites, detailState] : [detailState];
     localStorage.setItem('favorites', JSON.stringify(value));
   }
   
   const removeFromLocalStorage = (id: string) => {
     const favorites = JSON.parse(localStorage.getItem('favorites') ?? '[]');
-    const value = favorites.filter((obj: {url: string, id: string}) => obj.id === id)
+    const value = favorites.filter((photo: DetailState) => photo.id !== id)
     localStorage.setItem('favorites', JSON.stringify(value));
   }
-
 
   const value: CtxTypes = { imgDetail: detailState, dispatchImgDetail: dispatchDetail, addToLocalStorage,removeFromLocalStorage };
   return <ctx.Provider value={value}>{children}</ctx.Provider>;
